refactor(Site Pages): Moves site pages into pages folder.
